fix(*): some cases handling & UI updates
