+++ app/components/settings/onboarding/UserList.tsx
@@ -8,12 +8,17 @@ import {
   UserProjectListResponse,
   UserProjectDeleteResponse,
 } from "@/app/lib/types/onboarding";
-import { formatRelativeTime, isValidEmail } from "@/app/lib/utils";
-import { ArrowLeftIcon } from "@/app/components/icons";
-import { Button, Field } from "@/app/components";
+import { formatRelativeTime } from "@/app/lib/utils";
+import {
+  ArrowLeftIcon,
+  CheckCircleIcon,
+  TrashIcon,
+} from "@/app/components/icons";
+import { Button } from "@/app/components";
 import { useAuth } from "@/app/lib/context/AuthContext";
 import { apiFetch } from "@/app/lib/apiClient";
 import { useToast } from "@/app/components/Toast";
+import AddUserModal from "./AddUserModal";
 
 interface UserListProps {
   organization: Organization;
@@ -51,9 +56,7 @@ export default function UserList({
 
   const [users, setUsers] = useState<UserProject[]>([]);
   const [isLoading, setIsLoading] = useState(true);
-  const [userRows, setUserRows] = useState([{ email: "", full_name: "" }]);
-  const [rowErrors, setRowErrors] = useState<Record<number, string>>({});
-  const [isAdding, setIsAdding] = useState(false);
+  const [showAddModal, setShowAddModal] = useState(false);
   const [removingId, setRemovingId] = useState<number | null>(null);
 
   const fetchUsers = useCallback(async () => {
@@ -76,102 +79,13 @@ export default function UserList({
     fetchUsers();
   }, [fetchUsers]);
 
-  const updateRow = (
-    index: number,
-    field: "email" | "full_name",
-    value: string,
-  ) => {
-    setUserRows((prev) =>
-      prev.map((r, i) => (i === index ? { ...r, [field]: value } : r)),
-    );
-    if (rowErrors[index]) {
-      setRowErrors((prev) => {
-        const next = { ...prev };
-        delete next[index];
-        return next;
-      });
-    }
-  };
-
-  const addRow = () => {
-    setUserRows((prev) => [...prev, { email: "", full_name: "" }]);
-  };
-
-  const removeRow = (index: number) => {
-    if (userRows.length === 1) return;
-    setUserRows((prev) => prev.filter((_, i) => i !== index));
-    setRowErrors((prev) => {
-      const next: Record<number, string> = {};
-      Object.entries(prev).forEach(([k, v]) => {
-        const key = Number(k);
-        if (key < index) next[key] = v;
-        else if (key > index) next[key - 1] = v;
-      });
-      return next;
-    });
-  };
-
-  const handleAddUsers = async () => {
-    const filledRows = userRows.filter((r) => r.email.trim());
-
-    if (filledRows.length === 0) {
-      setRowErrors({ 0: "Please enter an email address" });
-      return;
-    }
-
-    const errors: Record<number, string> = {};
-    userRows.forEach((row, i) => {
-      if (row.email.trim() && !isValidEmail(row.email.trim())) {
-        errors[i] = "Invalid email address";
-      }
-    });
-
-    if (Object.keys(errors).length > 0) {
-      setRowErrors(errors);
-      return;
-    }
-
-    setRowErrors({});
-    setIsAdding(true);
-
-    try {
-      const payload = filledRows.map((r) => {
-        const entry: { email: string; full_name?: string } = {
-          email: r.email.trim(),
-        };
-        if (r.full_name.trim()) entry.full_name = r.full_name.trim();
-        return entry;
-      });
-
-      await apiFetch<UserProjectListResponse>("/api/user-projects", apiKey, {
-        method: "POST",
-        body: JSON.stringify({
-          organization_id: organization.id,
-          project_id: project.id,
-          users: payload,
-        }),
-      });
-      toast.success(
-        `${filledRows.length} user${filledRows.length > 1 ? "s" : ""} added successfully`,
-      );
-      setUserRows([{ email: "", full_name: "" }]);
-      await fetchUsers();
-    } catch (err) {
-      toast.error(err instanceof Error ? err.message : "Failed to add users");
-    } finally {
-      setIsAdding(false);
-    }
-  };
-
   const handleRemoveUser = async (userId: number) => {
     setRemovingId(userId);
     try {
       await apiFetch<UserProjectDeleteResponse>(
         `/api/user-projects/${userId}?project_id=${project.id}`,
         apiKey,
-        {
-          method: "DELETE",
-        },
+        { method: "DELETE" },
       );
       toast.success("User removed");
       setUsers((prev) => prev.filter((u) => u.user_id !== userId));
@@ -188,71 +102,24 @@ export default function UserList({
         <ArrowLeftIcon className="w-3.5 h-3.5" /> Back to projects
       </Button>
 
-      <div className="mb-4">
-        <h2 className="text-lg font-semibold text-text-primary">
-          {project.name}
-        </h2>
-        <p className="text-xs text-text-secondary mt-0.5">
-          {organization.name} &middot;{" "}
-          {isLoading
-            ? "Loading users..."
-            : `${users.length} user${users.length !== 1 ? "s" : ""}`}
-        </p>
-      </div>
-
-      {currentUser?.is_superuser && (
-        <div className="mb-6 p-4 rounded-lg border border-border bg-white">
-          <p className="text-sm font-medium text-text-primary mb-3">
-            Add Users
+      <div className="flex items-center justify-between mb-4">
+        <div>
+          <h2 className="text-lg font-semibold text-text-primary">
+            {project.name}
+          </h2>
+          <p className="text-xs text-text-secondary mt-0.5">
+            {organization.name} &middot;{" "}
+            {isLoading
+              ? "Loading users..."
+              : `${users.length} user${users.length !== 1 ? "s" : ""}`}
           </p>
-          <div className="space-y-3">
-            {userRows.map((row, index) => (
-              <div key={index} className="flex gap-2 items-start">
-                <div className="flex-1">
-                  <Field
-                    label=""
-                    value={row.email}
-                    onChange={(val) => updateRow(index, "email", val)}
-                    placeholder="Email address"
-                    error={rowErrors[index]}
-                  />
-                </div>
-                <div className="flex-1">
-                  <Field
-                    label=""
-                    value={row.full_name}
-                    onChange={(val) => updateRow(index, "full_name", val)}
-                    placeholder="Full name (optional)"
-                  />
-                </div>
-                {userRows.length > 1 && (
-                  <Button
-                    variant="ghost"
-                    size="sm"
-                    onClick={() => removeRow(index)}
-                  >
-                    &times;
-                  </Button>
-                )}
-              </div>
-            ))}
-          </div>
-          <div className="flex items-center justify-between mt-3">
-            <Button variant="ghost" size="sm" onClick={addRow}>
-              + Add another
-            </Button>
-            <Button
-              onClick={handleAddUsers}
-              disabled={isAdding || userRows.every((r) => !r.email.trim())}
-              size="md"
-            >
-              {isAdding
-                ? "Adding..."
-                : `Add User${userRows.filter((r) => r.email.trim()).length > 1 ? "s" : ""}`}
-            </Button>
-          </div>
         </div>
-      )}
+        {currentUser?.is_superuser && (
+          <Button size="sm" onClick={() => setShowAddModal(true)}>
+            + Add User
+          </Button>
+        )}
+      </div>
 
       {isLoading ? (
         <UserListSkeleton />
@@ -278,31 +145,41 @@ export default function UserList({
                   Added {formatRelativeTime(user.inserted_at)}
                 </p>
               </div>
-              <div className="flex items-center gap-3">
+              <div className="flex items-center gap-2">
                 <span
-                  className={`text-[10px] px-2 py-0.5 rounded-full font-medium ${
+                  className={`inline-flex items-center gap-1 text-xs px-2 py-1 rounded-full font-medium ${
                     user.is_active
                       ? "bg-green-50 text-green-700 border border-green-200"
                       : "bg-neutral-100 text-text-secondary border border-border"
                   }`}
                 >
+                  <CheckCircleIcon className="w-3.5 h-3.5" />
                   {user.is_active ? "Active" : "Inactive"}
                 </span>
                 {currentUser?.is_superuser && (
-                  <Button
-                    variant="danger"
-                    size="sm"
+                  <button
                     onClick={() => handleRemoveUser(user.user_id)}
                     disabled={removingId === user.user_id}
+                    className="inline-flex items-center gap-1.5 text-xs px-2.5 py-1.5 font-medium rounded-lg border border-red bg-white text-red-600 hover:bg-red-100 transition-colors cursor-pointer disabled:opacity-50"
                   >
+                    <TrashIcon className="w-3.5 h-3.5" />
                     {removingId === user.user_id ? "Removing..." : "Remove"}
-                  </Button>
+                  </button>
                 )}
               </div>
             </div>
           ))}
         </div>
       )}
+
+      <AddUserModal
+        open={showAddModal}
+        onClose={() => setShowAddModal(false)}
+        organizationId={organization.id}
+        projectId={project.id}
+        apiKey={apiKey}
+        onUsersAdded={fetchUsers}
+      />
     </div>
   );
 }
