+++ app/components/settings/onboarding/index.ts
@@ -4,3 +4,4 @@ export { default as OrganizationList } from "./OrganizationList";
 export { default as ProjectList } from "./ProjectList";
 export { default as StepIndicator } from "./StepIndicator";
 export { default as UserList } from "./UserList";
+export { default as AddUserModal } from "./AddUserModal";
